% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_both.R
\name{do_both}
\alias{do_both}
\title{do_both()}
\usage{
do_both(interactive = FALSE, ...)
}
\arguments{
\item{interactive}{A logical. Indicates whether user wants to input
importance of different budgets aspects.}

\item{...}{Other arguments passed on to layer(). These are often aesthetics,
used to set an aesthetic to a fixed value, like colour = "red" or size = 3.
They may also be parameters to the paired geom/stat.}
}
\value{
The output of \link{monthly_budget_more} using the inputed income as the income
argument
}
\description{
Calls \link{interact} and \link{monthly_budget_more} at the same time making the output
of \link{interact} the argument in \link{monthly_budget_more}
}
\examples{
#inc = 4000
#do_both()

}
