% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/food.R
\name{food_budget}
\alias{food_budget}
\title{Food Budget}
\usage{
food_budget(..., pct = 0.15, tip = FALSE, food_split = FALSE)
}
\arguments{
\item{...}{Numeric vectors. The incomes for which the user wants to get food budgets.}

\item{pct}{A numeric vector. The percentage of income contributed to food. By default set to 0.15, the recommended percentage.}

\item{tip}{A logical parameter set to FALSE by default. If TRUE, function outputs budgeting tips.}

\item{food_split}{A logical parameter set to FALSE by default. If TRUE, function splits up the food budget and calculates how much of income should be spent on groceries and takeout.}
}
\value{
This returns the recommended food budget.
}
\description{
This function takes a percentage of the inputted monthly income and allocates it to their food budget.
}
\details{
Similar to \link{housing_budget}, \link{transportation_budget}, \link{healthcare_budget}, \link{savings_budget}, and \link{other_budget}
}
\examples{
# example data
example_income = 3000

# function
result = food_budget(example_income)
}
