% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/healthcare.R
\name{healthcare_budget}
\alias{healthcare_budget}
\title{Healthcare budget}
\usage{
healthcare_budget(..., pct = 0.1)
}
\arguments{
\item{...}{Numeric vectors. The incomes for which the user wants to get healthcare budgets.}

\item{pct}{A numeric vector. The percentage of income contributed to healthcare. By default set to 0.1, the recommended percentage.}
}
\value{
Returns the recommended healthcare budget
}
\description{
This function takes a percentage of the inputted monthly income and allocates it to their housing budget.
}
\details{
Similar to \link{housing_budget}, \link{food_budget}, \link{transportation_budget}, \link{savings_budget}, and \link{other_budget}
}
\examples{
# create example data
example_income = 4000

# Compute healthcare budget
result = healthcare_budget(example_income)

}
