% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housing.R
\name{housing_budget}
\alias{housing_budget}
\title{Housing Budget}
\usage{
housing_budget(..., pct = 0.3, tip = FALSE)
}
\arguments{
\item{...}{Numeric vectors. The incomes for which the user wants to get housing budgets.}

\item{pct}{A numeric vector. The percentage of income contributed to housing. By default set to 0.3, the recommended percentage.}

\item{tip}{A logical parameter set to FALSE by default. If TRUE, function outputs budgeting tips.}
}
\value{
This returns the recommended housing budget.
}
\description{
This function takes a percentage of the inputted monthly income and allocates it to their housing budget.
}
\details{
Similar to \link{food_budget}, \link{transportation_budget}, \link{healthcare_budget}, \link{savings_budget}, and \link{other_budget}
}
\examples{
# example data
example_income = 3000

# function
result = housing_budget(example_income)
}
