% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housing_recommendation.R
\name{housing_rec}
\alias{housing_rec}
\title{Housing Recommendations}
\usage{
housing_rec(...)
}
\arguments{
\item{...}{These can be either numeric or character vectors. If numeric, then it must be the user's annual income. If a character, then the input must either be the abbreviation of the state for which they want to buy a house, or one of "Rent", "Conventional", "Contemporary", or "Multi-Storey Home", reflecting the kind of home they want to own.}
}
\value{
Returns recommendations
}
\description{
This function will output reccomendations based on the user's inputted income, state, and type of desired housing.
}
\details{
This function builds on the \link{housing_budget} function.
}
\examples{

library(betterBudgets)

# Get a housing recommendation for a user living in MA, a user who earns $100,000
# annually, and a user who wants to own a multi-storey home

housing_rec("MA", 100000, "Multi-Storey Home")

}
