% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthly.R
\name{interact}
\alias{interact}
\title{Interact}
\usage{
interact()
}
\value{
Returns the income entered by the user
}
\description{
A helper function that creates an income variable created by requesting monthly input from the user
}
\examples{
# Get budget table and distribution for income and color the graph blue
# monthly_budget_more(fill = "cornflowerblue", color = "black")

}
