% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthly.R
\name{monthly_budget_more}
\alias{monthly_budget_more}
\title{Monthly Budget}
\usage{
monthly_budget_more(income, interactive = FALSE, ...)
}
\arguments{
\item{income}{User's monthly income. A numeric value.}

\item{interactive}{A logical value. Indicates whether user wants to input importance of different budgets aspects.}

\item{...}{Other arguments passed on to layer(). These are often aesthetics, used to set an aesthetic to a fixed value, like colour = "red" or size = 3. They may also be parameters to the paired geom/stat.}
}
\value{
A table of budget distribution and visualization of the same.
}
\description{
A function that takes the income and outputs a table and graph of the budget distribution.
}
\details{
Uses the \link{food_budget}, \link{housing_budget}, \link{savings_budget}, \link{transportation_budget}, \link{healthcare_budget}, and \link{other_budget} functions.
}
\examples{
# example_income = 10000
# Get budget table and distribution for income and color the graph blue
# monthly_budget_more(income = example_income, fill = "cornflowerblue", color = "black")

}
