% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other.R
\name{other_budget}
\alias{other_budget}
\title{Other budget}
\usage{
other_budget(..., pct = 0.1)
}
\arguments{
\item{...}{Numeric vectors. The incomes for which the user wants to get other/miscellaneous budgets.}

\item{pct}{A numeric vector. The percentage of income contributed to other. By default set to 0.1, the recommended percentage.}
}
\value{
Returns the recommended savings budget
}
\description{
The function takes a percentage of the inputted monthly income and allocates it to their budget for other/miscellaneous expenses that are not included in the housing, food, transportation, healthcare, or savings budgets.
}
\details{
Similar to \link{housing_budget}, \link{food_budget}, \link{transportation_budget}, \link{healthcare_budget}, and \link{savings_budget}
}
\examples{
# create example data
example_income = 4000

# Compute other budget
result = other_budget(example_income)

}
