% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/food_class.R
\name{tips.food}
\alias{tips.food}
\title{tips.food}
\usage{
\method{tips}{food}(x)
}
\arguments{
\item{x}{An object used to select a method}
}
\value{
Returns tips to help the user save money on food.
}
\description{
Tips method for the food class. This method prints out budgeting tips for food.
}
\details{
Similar to \link{tips.housing} and \link{tips.transportation}
}
\examples{
tips.food()
}
