% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housing_class.R
\name{tips.housing}
\alias{tips.housing}
\title{tips.housing}
\usage{
\method{tips}{housing}(x)
}
\arguments{
\item{x}{An object used to select a method}
}
\value{
Returns tips to help the user save money on housing.
}
\description{
Tips method for the housing class. This method prints out budgeting tips for housing.
}
\details{
Similar to \link{tips.food} and \link{tips.transportation}
}
\examples{
tips.housing()
}
