% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tips.R
\name{tips}
\alias{tips}
\title{tips}
\usage{
tips(x)
}
\arguments{
\item{x}{An object used to select a method}
}
\value{
Returns tips
}
\description{
A generic function in which new tips methods can be easily added for new classes.
}
\examples{
tips.transportation()
}
