% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transportation_class.R
\name{tips.transportation}
\alias{tips.transportation}
\title{tips.transportation}
\usage{
\method{tips}{transportation}(x)
}
\arguments{
\item{x}{An object used to select a method}
}
\value{
Returns tips to help the user save money on transportation.
}
\description{
Tips method for the transportation class. This method prints out budgeting tips for transportation.
}
\details{
Similar to \link{tips.food} and \link{tips.housing}
}
\examples{
tips.transportation()
}
