% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transportation.R
\name{transportation_budget}
\alias{transportation_budget}
\title{Transportation Budget}
\usage{
transportation_budget(
  income,
  pct = 0.15,
  tip = FALSE,
  rec = FALSE,
  dist = 25,
  consump = 25,
  cpg = 3
)
}
\arguments{
\item{income}{The user's inputted monthly income value.}

\item{pct}{A numeric vector. The percentage of income contributed to transportation. By default set to 0.15, the recommended percentage.}

\item{tip}{A logical parameter set to FALSE by default. If TRUE, function outputs budgeting tips.}

\item{rec}{A logical parameter set to FALSE by default. If TRUE, the function will output recommendations to save the user money.}

\item{dist}{The distance the user travels in a day. Set to 25 miles by default.}

\item{consump}{The user's vehicle consumption. Set to 25 miles by default.}

\item{cpg}{The fuel's price per gallon. Set to $3 by default.}
}
\value{
This returns the recommended transportation budget.
}
\description{
This function takes a percentage of the inputted monthly income and allocates it to their transportation budget.
}
\details{
Similar to \link{housing_budget}, \link{food_budget}, \link{healthcare_budget}, \link{savings_budget}, and \link{other_budget}
}
\examples{
# example data
example_income = 3000

# function
result = transportation_budget(example_income)
}
