% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_income.R
\name{user_income}
\alias{user_income}
\title{User Income}
\usage{
user_income(income)
}
\arguments{
\item{income}{Numeric value greater than 0. The user's yearly income.}
}
\value{
Numeric value. The user's monthly income.
}
\description{
This function takes a user's annual income then calculates their monthly income by dividing it by 12
}
\examples{
# create a dummy data
user_income(2000)

# take user's annual income as input
}
